Remove required and optional fields
