+++ app/lib/grades/pages/grades_dialog/grades_dialog.dart
@@ -17,8 +17,7 @@ import 'package:sharezone/grades/pages/grades_dialog/grades_dialog_view.dart';
 import 'package:sharezone/grades/pages/shared/saved_grade_icons.dart';
 import 'package:sharezone_widgets/sharezone_widgets.dart';
 
-part 'optional_fields.dart';
-part 'required_fields.dart';
+part 'fields.dart';
 
 class GradesDialog extends StatelessWidget {
   const GradesDialog({super.key});
@@ -35,25 +34,15 @@ class GradesDialog extends StatelessWidget {
         ),
       ),
       builder: (context, _) {
-        final view = context.watch<GradesDialogController>().view;
         return Scaffold(
           appBar: AppBar(
             actions: const [_SaveButton()],
           ),
-          body: SingleChildScrollView(
-            padding: const EdgeInsets.all(8),
+          body: const SingleChildScrollView(
+            padding: EdgeInsets.all(8),
             child: SafeArea(
               child: MaxWidthConstraintBox(
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    const SizedBox(height: 4),
-                    const _RequiredFields(),
-                    const Divider(),
-                    const SizedBox(height: 4),
-                    _OptionalFields(view: view),
-                  ],
-                ),
+                child: _Fields(),
               ),
             ),
           ),
@@ -83,14 +72,8 @@ class _SaveButton extends StatelessWidget {
   }
 }
 
-class _Section extends StatelessWidget {
-  const _Section({
-    required this.title,
-    required this.children,
-  });
-
-  final String title;
-  final List<Widget> children;
+class _Fields extends StatelessWidget {
+  const _Fields();
 
   @override
   Widget build(BuildContext context) {
@@ -100,35 +83,19 @@ class _Section extends StatelessWidget {
               color: Theme.of(context).colorScheme.onSurface.withOpacity(0.5),
             ),
           ),
-      child: Column(
+      child: const Column(
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [
-          _SectionTitle(title: title),
-          const SizedBox(height: 8),
-          for (final child in children)
-            Padding(
-              padding: const EdgeInsets.only(bottom: 4),
-              child: child,
-            ),
+          _GradeValue(),
+          _GradingSystem(),
+          _Subject(),
+          _Date(),
+          _GradingType(),
+          _Term(),
+          _IntegrateGradeIntoSubjectGrade(),
+          _Title(),
+          _Details(),
         ],
-      ),
-    );
-  }
-}
-
-class _SectionTitle extends StatelessWidget {
-  const _SectionTitle({
-    required this.title,
-  });
-
-  final String title;
-
-  @override
-  Widget build(BuildContext context) {
-    return Text(
-      title.toUpperCase(),
-      style: TextStyle(
-        color: Theme.of(context).colorScheme.onSurface.withOpacity(0.5),
       ),
     );
   }
