+++ app/test_goldens/grades/pages/create_term_page/create_term_page_test.mocks.dart
@@ -12,7 +12,7 @@ import 'package:fast_immutable_collections/fast_immutable_collections.dart'
 import 'package:flutter/foundation.dart' as _i11;
 import 'package:mockito/mockito.dart' as _i1;
 import 'package:mockito/src/dummies.dart' as _i8;
-import 'package:rxdart/subjects.dart' as _i2;
+import 'package:rxdart/rxdart.dart' as _i2;
 import 'package:sharezone/grades/grades_service/grades_service.dart' as _i4;
 import 'package:sharezone/grades/models/grade_id.dart' as _i7;
 import 'package:sharezone/grades/models/subject_id.dart' as _i6;
@@ -306,9 +306,9 @@ class MockGradesService extends _i1.Mock implements _i4.GradesService {
         ),
       ) as _i3.IList<_i4.GradeType>);
   @override
-  void createCustomGradeType(_i4.GradeType? gradeType) => super.noSuchMethod(
+  void addCustomGradeType(_i4.GradeType? gradeType) => super.noSuchMethod(
         Invocation.method(
-          #createCustomGradeType,
+          #addCustomGradeType,
           [gradeType],
         ),
         returnValueForMissingStub: null,
